Upgrade version to 2.21.0 (#409)
* Upgraded GeoServer from version 2.20.4 to 2.21.0

+++ scripts/functions.sh
@@ -200,14 +200,17 @@ function install_plugin() {
   fi
   EXT=$2
 
-  unzip "${DATA_PATH}"/"${EXT}".zip -d /tmp/gs_plugin
-  if [[ -f /geoserver/start.jar ]]; then
-    cp -r -u -p /tmp/gs_plugin/*.jar /geoserver/webapps/geoserver/WEB-INF/lib/
+  if [[ -f "${DATA_PATH}"/"${EXT}".zip ]];then
+     unzip "${DATA_PATH}"/"${EXT}".zip -d /tmp/gs_plugin
+     if [[ -f /geoserver/start.jar ]]; then
+       cp -r -u -p /tmp/gs_plugin/*.jar /geoserver/webapps/geoserver/WEB-INF/lib/
+     else
+       cp -r -u -p /tmp/gs_plugin/*.jar "${CATALINA_HOME}"/webapps/geoserver/WEB-INF/lib/
+     fi
+     rm -rf /tmp/gs_plugin
   else
-    cp -r -u -p /tmp/gs_plugin/*.jar "${CATALINA_HOME}"/webapps/geoserver/WEB-INF/lib/
-  fi
-  rm -rf /tmp/gs_plugin
-
+    echo -e "\e[32m ${EXT} extension will not be installed because it is not available \033[0m"
+ fi
 }
 
 # Helper function to setup disk quota configs and database configurations
@@ -269,7 +272,7 @@ function geoserver_logging() {
   if [[ ! -f ${GEOSERVER_DATA_DIR}/logging.xml ]];then
     echo "
 <logging>
-  <level>${GEOSERVER_LOG_LEVEL}.properties</level>
+  <level>${GEOSERVER_LOG_LEVEL}</level>
   <location>logs/geoserver.log</location>
   <stdOutLogging>true</stdOutLogging>
 </logging>
